refactor language toggle service and add code comments
