+++ component-library/component-lib/src/lib/header-footer/ircc-ds-angular-header-footer.module.ts
@@ -2,12 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { HeaderComponent } from './header/header.component';
 import { FooterComponent } from './footer/footer.component';
-import { LanguageSwitchComponent } from './language-switch/language-switch.component';
 import { IrccDsAngularComponentsSharedModule } from '../shared/ircc-ds-angular-component-shared.module';
 import { TranslateModule } from '@ngx-translate/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HiddenNavComponent } from './hidden-nav/hidden-nav.component';
 import { ThemeSwitchComponent } from './theme-switch/theme-switch.component';
+import { LanguageSwitchComponent } from './language-switch/language-switch.component';
+
 
 const IrccDsHeaderFooterComponents = [
   HeaderComponent,
@@ -28,4 +29,4 @@ const IrccDsHeaderFooterComponents = [
   ],
   exports: [...IrccDsHeaderFooterComponents]
 })
-export class IrccDsAngularHeaderFooterModule {}
+export class IrccDsAngularHeaderFooterModule { }
