--- component-library/component-lib/src/lib/header-footer/language-switch/language-switch-button.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { LanguageSwitchButtonService } from './language-switch-button.service';
-
-describe('LanguageSwitchService', () => {
-  let service: LanguageSwitchButtonService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(LanguageSwitchButtonService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
