+++ component-library/src/app/shell/shell.component.html
@@ -2,7 +2,7 @@
   <div class="header">
     <ircc-cl-lib-header
       [id]="'header'"
-      [themeToggle]="true"
+      [themeToggle]="false"
     >
       <div
         class="container"
